refactoring for devex
